rename: rename package Model to VO
모델 패키지 명을 VO로 변경
