+++ app/src/main/java/campus/tech/kakao/map/Presenter/View/Adapter/SearchResultAdapter.kt
@@ -1,16 +1,14 @@
 package campus.tech.kakao.map.Presenter.View.Adapter
 
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import campus.tech.kakao.map.Domain.Model.Place
+import campus.tech.kakao.map.Domain.VO.Place
 import campus.tech.kakao.map.R
-import campus.tech.kakao.map.Domain.Model.PlaceCategory
+import campus.tech.kakao.map.Domain.VO.PlaceCategory
 
 class SearchResultAdapter(
     val onClickAdd: (id: Int) -> Unit
